Quote string when resource is query
Prints the query as a quoted string when adding it as a resource to the
trace.

The Datadog tracer prints traces to stdout when the environment
variable DD_FLUSH_TO_LOG is set to true. If the resource field contains
newline characters this ends up printing the trace over multiple lines,
which then causes the Datadog log forwarder to not be able to send the
traces to Datadog.