Support full ARN in secretsmanagerauth
If your service key name happens to end in a hyphen and six characters,
AWS will assume this to be a full ARN when fetching the secret, and the
policy on the secret in the users account will deny the request.

By fetching the full ARN and using this in the config this problem can
be circumvented.

This should be implemented in way where if an ARN is configured the
value is used, otherwise the legacy behavior is assumed.