Support logging only on only (#138)
Adds a function which can be used in instances where you only want to
log when there's an active trace. The function returns a noop logger
if a trace is not found in the context, which would then not print
anything when logged.

This would be useful in instances where an application is configured to
trace less than 100% of requests, and the application wants to avoid
logging e.g. debug statements when no trace is present.