Add aws-sdk-go-v2 middleware for SQS/SNS
Adds a middleware which can be used to inject trace information in
message attributes on SNS & SQS messages. Uses the Inject functionality
of the DataDog tracing library so that it respects
DD_PROPAGATION_STYLE_INJECT settings in the environment. I.e. if this
configuration is set to DataDog,B3 you'll get both DataDog and B3 header
injected in the requests.

Adds simpler functions for starting traces directly from the SQS event
received by a Lambda function.

See example modules for how to use the library.

Note: Naming follows the module named of the AWS SDK v2 for Go, similar
to how the contrib folder of the DataDog tracing library is structured.